add yasg, check http method

+++ todolists/views.py
@@ -1,45 +1,48 @@
 from django.shortcuts import render, redirect, get_object_or_404
-from django.http import JsonResponse
+
+from rest_framework.decorators import api_view
+from rest_framework.response import Response
+import json
 
 from .models import TodoList
 from .serializers import TodoListSerializer
 
-def alltodos(request):
+
+@api_view(['GET'])
+def read(request):
     todolist = TodoList.objects.order_by('pk')
 
-    todos = TodoListSerializer(todolist, many=True)
-    
+    serializer = TodoListSerializer(todolist, many=True)
+
     context = {
-        'todos': todos.data,
+        'todos': serializer.data
     }
-
-    return JsonResponse(context)
-
-def addtodo(request):
-
-    # print(request.GET)
-    todo = TodoListSerializer(data=request.GET)
-    if todo.is_valid(raise_exception=True):
-        res = todo.save()
-
-    return redirect('todolists:all')
+    return Response(context)
 
 
-def update(request, todo_pk):
+@api_view(['PUT', 'POST', 'DELETE'])
+def cud(request, todo_pk):
     todo = get_object_or_404(TodoList, pk=todo_pk)
+    post_data = json.loads(request.body.decode('utf-8'))
 
-    serializer = TodoListSerializer(data=request.GET, instance=todo)
-    
-    if serializer.is_valid(raise_exception=True):
-        serializer.save()
-        todo = get_object_or_404(TodoList, pk=todo_pk)
-    
-    return redirect('todolists:all')
+    if request.method == "POST":
+        serializer = TodoListSerializer(data=post_data)
 
+        if serializer.is_valid(raise_exception=True):
+            serializer.save()
 
-def delete(request, todo_pk):
-    todo = get_object_or_404(TodoList, pk=todo_pk)
-    
-    todo.delete()
+    elif request.method == "PUT":
+        serializer = TodoListSerializer(data=post_data, instance=todo)
+        
+        if serializer.is_valid(raise_exception=True):
+            serializer.save()
 
-    return redirect('todolists:all')
+    elif request.method == "DELETE":
+        todo.delete()
+        
+    todolist = TodoList.objects.order_by('pk')
+    todos = TodoListSerializer(todolist, many=True)
+    context = {
+        'todos': todos.data
+    }
+    return Response(context)
